pyproject.toml
[project]
name = "andera"
version = "0.1.0"
description = "General Browser Agent for audit-evidence collection (Andera work trial)"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "cryptography>=42",
    "litellm>=1.48",
    "playwright>=1.47",
    "langgraph>=0.2",
    "langgraph-checkpoint-sqlite>=2.0",
    "typer>=0.12",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "jsonschema>=4.20",
    "anthropic>=0.96.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "fakeredis>=2.20",
]
redis = [
    "redis>=5.0",
]
postgres = [
    "asyncpg>=0.29",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary]>=3.1",
]

[project.scripts]
andera = "andera.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/andera"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
